# openapi.yaml
openapi: 3.0.1
info:
  title: Kramesh Portfolio Backend API
  description: API for managing portfolio content (education, experience, skills, about, projects, blogs, contact, certifications, auth).
  version: "1.0"
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /about:
    get:
      tags:
        - about
      operationId: getAbout
      summary: Get the About record
      description: Returns the single About-page record (name, bio, links).
      responses:
        200:
          description: About record found and returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/About"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/about_get_server_error.json
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
    post:
      tags:
        - about
      operationId: createAbout
      summary: Create the About record
      description: Creates the About-page record (multipart form, optional image file). Requires admin authentication.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AboutFields"
      responses:
        201:
          description: About record created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/About"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/about_post_unauthorized.json
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/about_post_server_error.json
        400:
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
    put:
      tags:
        - about
      operationId: updateAbout
      summary: Update the About record
      description: Updates (or upserts) the About-page record (multipart form, optional image file). Requires admin authentication.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AboutFields"
      responses:
        200:
          description: About record updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/About"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/about_put_unauthorized.json
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/about_put_server_error.json
        400:
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
...
  /blogs:
    get:
      tags:
        - blogs
      operationId: getBlogs
      summary: List published blogs
      description: Returns an array of published blog posts, newest first.
      responses:
        200:
          description: Blogs found and returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Blog"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/blogs_get_server_error.json
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
    post:
      tags:
        - blogs
      operationId: createBlog
      summary: Create a blog post
      description: Creates a new blog post. Requires admin authentication.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogFields"
      responses:
        201:
          description: Blog created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
        400:
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/blogs_post_unauthorized.json
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/blogs_post_server_error.json
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
  /blogs/all:
    get:
      tags:
        - blogs
      operationId: getAllBlogs
      summary: List all blogs (admin)
      description: Returns all blogs, including unpublished. Requires admin authentication.
      security:
        - bearerAuth: []
      responses:
        200:
          description: All blogs found and returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Blog"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/blogs_all_unauthorized.json
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/blogs_all_server_error.json
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
  /blogs/{slug}:
    get:
      tags:
        - blogs
      operationId: getBlogBySlug
      summary: Get a blog by slug
      description: Returns a single blog post matching the given slug.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          example: my-first-blog-post
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
      responses:
        200:
          description: Blog found and returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
        404:
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not_found:
                  value:
                    message: Blog not found
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/blogs_slug_server_error.json
  /blogs/{id}:
    put:
      tags:
        - blogs
      operationId: updateBlog
      summary: Update a blog by ID
      description: Updates an existing blog post. Requires admin authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
          example: 64f1a2b3c4d5e6f7a8b9c0d5
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogFields"
      responses:
        200:
          description: Blog updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
        400:
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/blogs_put_unauthorized.json
        404:
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/blogs_put_server_error.json
    delete:
      tags:
        - blogs
      operationId: deleteBlog
      summary: Delete a blog by ID
      description: Deletes a blog post. Requires admin authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
          example: 64f1a2b3c4d5e6f7a8b9c0d5
      responses:
        200:
          description: Blog deleted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteMessage"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/blogs_delete_unauthorized.json
        404:
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      title: Error response
      type: object
      properties:
        message:
          type: string
          example: Not authorized
      required:
        - message
    DeleteMessage:
      title: Delete confirmation message
      type: object
      properties:
        message:
          type: string
          example: Education deleted
      required:
        - message
...
    AboutFields:
      title: About fields
      description: Editable fields of the About record.
      type: object
      properties:
        name:
          type: string
          example: Kuruba Ramesh
        title:
          type: string
          example: Full Stack Developer
        bio:
          type: string
          example: I build web applications using the MERN stack and Java/Spring Boot.
        image_url:
          type: string
          example: https://res.cloudinary.com/demo/image/upload/about.jpg
        resume_url:
          type: string
          example: https://res.cloudinary.com/demo/raw/upload/resume.pdf
        github:
          type: string
          example: https://github.com/KRameshr
        linkedin:
          type: string
          example: https://linkedin.com/in/kurubaramesh
        email:
          type: string
          example: ramesh@example.com
      required:
        - name
        - title
        - bio
    About:
      title: About
      description: The About-page record.
      allOf:
        - $ref: "#/components/schemas/AboutFields"
        - type: object
          properties:
            _id:
              type: string
              pattern: ^[0-9a-fA-F]{24}$
              example: 64f1a2b3c4d5e6f7a8b9c0d8
            image_public_id:
              type: string
              example: about_abc123
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            __v:
              type: integer
              example: 0
          required:
            - _id
...
    BlogFields:
      title: Blog fields
      description: Editable fields of a blog post.
      type: object
      properties:
        title:
          type: string
          example: How I Set Up Specmatic Contract Testing
        slug:
          type: string
          example: how-i-set-up-specmatic-contract-testing
        content:
          type: string
          example: In this post I walk through setting up Specmatic contract tests...
        excerpt:
          type: string
          example: A step-by-step guide to contract testing with Specmatic.
        cover_image_url:
          type: string
          example: https://res.cloudinary.com/demo/image/upload/blog.jpg
        is_published:
          type: boolean
          example: true
      required:
        - title
        - slug
        - content
    Blog:
      title: Blog
      description: A blog post.
      allOf:
        - $ref: "#/components/schemas/BlogFields"
        - type: object
          properties:
            _id:
              type: string
              pattern: ^[0-9a-fA-F]{24}$
              example: 64f1a2b3c4d5e6f7a8b9c0d5
            cover_image_public_id:
              type: string
              example: blog_abc123
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            __v:
              type: integer
              example: 0
          required:
            - _id
...